spec/style cleanup; implement locale
