+++ lib/csl/styles/tex-citealts.csl
@@ -4,7 +4,6 @@
     <title>TeX style for macro citealt*</title>
     <id>tex-citetalts</id>
   </info>
-  <locale xml:lang="en"/>
   <macro name="author-long">
     <names variable="author">
       <name form="short" and="text" initialize="false" sort-separator=""/>
