+++ samples/tex/sample-authoryear.html
@@ -527,7 +527,7 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 </div>
 <div id="footer">
 <div id="footer-text">
-Last updated 2017-09-18 21:52:10 CEST
+Last updated 2017-09-24 13:09:54 CEST
 </div>
 </div>
 </body>
